[DAD-550] refactor: scrapService에서 각각의 서비스에 해당하는 부분을 해당 서비스로 이동
