+++ src/main/java/com/forever/dadamda/service/scrap/ArticleService.java
@@ -1,6 +1,7 @@
 package com.forever.dadamda.service.scrap;
 
 import com.forever.dadamda.dto.ErrorCode;
+import com.forever.dadamda.dto.scrap.GetArticleResponse;
 import com.forever.dadamda.dto.scrap.UpdateScrapRequest;
 import com.forever.dadamda.entity.scrap.Article;
 import com.forever.dadamda.entity.user.User;
@@ -9,6 +10,10 @@ import com.forever.dadamda.repository.ArticleRepository;
 import com.forever.dadamda.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import net.minidev.json.JSONObject;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -42,14 +47,12 @@ public class ArticleService {
     @Transactional
     public Article updateArticle(User user, UpdateScrapRequest updateScrapRequest) {
         Article article = articleRepository.findByIdAndUserAndDeletedDateIsNull(
-                updateScrapRequest.getScrapId(), user).orElseThrow(
-                () -> new NotFoundException(ErrorCode.NOT_EXISTS_SCRAP)
-        );
+                        updateScrapRequest.getScrapId(), user)
+                .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_EXISTS_SCRAP));
         article.update(updateScrapRequest.getTitle(), updateScrapRequest.getDescription(),
                 updateScrapRequest.getSiteName());
         article.updateArticle(updateScrapRequest.getAuthor(),
-                updateScrapRequest.getAuthorImageUrl(),
-                updateScrapRequest.getPublishedDate(),
+                updateScrapRequest.getAuthorImageUrl(), updateScrapRequest.getPublishedDate(),
                 updateScrapRequest.getBlogName());
 
         return article;
@@ -59,5 +62,19 @@ public class ArticleService {
     public Long getArticleCount(String email) {
         User user = userService.validateUser(email);
         return articleRepository.countByUserAndDeletedDateIsNull(user);
+    }
+
+    @Transactional
+    public Slice<GetArticleResponse> getArticles(String email, Pageable pageable) {
+        User user = userService.validateUser(email);
+
+        Sort sort = Sort.by(Sort.Direction.DESC, "createdDate");
+        PageRequest pageRequest = PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(),
+                sort);
+
+        Slice<Article> articleSlice = articleRepository.findAllByUserAndDeletedDateIsNull(user,
+                pageRequest).orElseThrow(() -> new NotFoundException(ErrorCode.NOT_EXISTS_SCRAP));
+
+        return articleSlice.map(GetArticleResponse::of);
     }
 }
