+++ src/main/java/com/forever/dadamda/service/scrap/ProductService.java
@@ -1,6 +1,7 @@
 package com.forever.dadamda.service.scrap;
 
 import com.forever.dadamda.dto.ErrorCode;
+import com.forever.dadamda.dto.scrap.GetProductResponse;
 import com.forever.dadamda.dto.scrap.UpdateScrapRequest;
 import com.forever.dadamda.entity.scrap.Product;
 import com.forever.dadamda.entity.user.User;
@@ -9,6 +10,10 @@ import com.forever.dadamda.repository.ProductRepository;
 import com.forever.dadamda.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import net.minidev.json.JSONObject;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -45,5 +50,19 @@ public class ProductService {
     public Long getProductCount(String email) {
         User user = userService.validateUser(email);
         return productRepository.countByUserAndDeletedDateIsNull(user);
+    }
+
+    @Transactional
+    public Slice<GetProductResponse> getProducts(String email, Pageable pageable) {
+        User user = userService.validateUser(email);
+
+        Sort sort = Sort.by(Sort.Direction.DESC, "createdDate");
+        PageRequest pageRequest = PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(),
+                sort);
+
+        Slice<Product> scrapSlice = productRepository.findAllByUserAndDeletedDateIsNull(user,
+                pageRequest).orElseThrow(() -> new NotFoundException(ErrorCode.NOT_EXISTS_SCRAP));
+
+        return scrapSlice.map(GetProductResponse::of);
     }
 }
